Revert "merge release 1.0.3"
