+++ CONTRIBUTING.md
@@ -40,18 +40,6 @@ Go to [issues page](https://github.com/FISCO-BCOS/console/issues)
 7. Wait the community to review the code
 8. Merged !!!!
 
-## Code formatting
-
-The code formatting tool are described by the [google-java-format-gradle-plugin](https://github.com/sherter/google-java-format-gradle-plugin).
-
-Execute the task `googleJavaFormat` to format all *.java files in the project
-```
-./gradlew goJF
-```
-Execute the task `verifyGoogleJavaFormat` to verify that all *.java files are formatted properly
-```          
-./gradlew verGJF
-```
 
 ## Continous integration
 
@@ -62,6 +50,6 @@ Execute the task `verifyGoogleJavaFormat` to verify that all *.java files are fo
 
 **Code quality**
 
-* Codacy: [![Codacy Badge](https://api.codacy.com/project/badge/Grade/a2a6c2eb499e42739d066ff775d1b288)](https://www.codacy.com/app/fisco/console?utm_source=github.com&amp;utm_medium=referral&amp;utm_content=FISCO-BCOS/console&amp;utm_campaign=Badge_Grade)
+* [![Codacy Badge](https://api.codacy.com/project/badge/Grade/a2a6c2eb499e42739d066ff775d1b288)](https://www.codacy.com/app/fisco/console?utm_source=github.com&amp;utm_medium=referral&amp;utm_content=FISCO-BCOS/console&amp;utm_campaign=Badge_Grade)
