+++ build.gradle
@@ -1,6 +1,3 @@
-plugins {
-  id 'com.github.sherter.google-java-format' version '0.8'
-}
 apply plugin: 'maven'
 apply plugin: 'java'
 apply plugin: 'eclipse'
@@ -18,45 +15,58 @@ repositories {
     mavenCentral()
 }
 
-googleJavaFormat {
-  options style: 'AOSP'
-}
+List alibaba = [
+        'com.alibaba:druid:1.0.29',
+        'com.alibaba:fastjson:1.2.29'
+]
 
-check.dependsOn.remove(test)
-check.dependsOn.remove(verifyGoogleJavaFormat)
+List logger = [
+	'org.slf4j:slf4j-log4j12:1.7.25'
+]
+
+def spring_version="4.1.8.RELEASE"
+List spring =[ 
+	"org.springframework:spring-core:$spring_version", 
+	"org.springframework:spring-beans:$spring_version", 
+	"org.springframework:spring-context:$spring_version", 
+	"org.springframework:spring-tx:$spring_version", 
+	"org.springframework:spring-jdbc:$spring_version", 
+	"org.springframework:spring-test:$spring_version"
+]
 
 // In this section you declare the dependencies for your production and test code
 dependencies {
-    compile group: "org.fisco-bcos", name: "web3sdk", version: "2.0.31-SNAPSHOT", changing: true
-//	compile group: 'org.fisco-bcos', name: 'web3sdk', version: '2.0.0-rc1'
-    compile group: 'org.slf4j', name:'slf4j-log4j12', version:'1.7.25'
-    compile group: 'org.jline', name: 'jline', version: '3.9.0'
-    compile group: 'io.bretty', name: 'console-table-builder', version: '1.2'
-    compile group: 'com.github.jsqlparser', name: 'jsqlparser', version: '2.0'
-    testCompile group: 'junit', name: 'junit', version: '4.1'
+	compile logger,spring
+	runtime logger,spring
+    
+	compile group:"org.fisco-bcos", name:"web3sdk", version:"2.0.3-SNAPSHOT", changing: true
+//	compile	'org.fisco-bcos:web3sdk:2.0.0-rc1'
+	compile 'org.jline:jline:3.10.0'
+	compile 'io.bretty:console-table-builder:1.2'
+	testCompile 'junit:junit:4.12'
 }
 
 configurations.all {
-    resolutionStrategy.cacheChangingModulesFor 0, 'seconds'
+	resolutionStrategy.cacheChangingModulesFor 0, 'seconds'
 }
 sourceSets {
-    main {
-        java {
-            srcDir 'src/main/java'
-        }
-        resources {
+	main {
+		java {
+	        srcDir 'src/main/java'
+		}
+        resources  {
             srcDir 'src/main/resources'
         }
-    }
+	}
 }
 
 jar {
-    destinationDir file('dist/apps')
-    archiveName project.name + '.jar'
-    exclude '**/*.xml'
-    exclude '**/*.properties'
-    exclude '**/*.crt'
-    exclude '**/*.key'
+	destinationDir file('dist/apps')
+	archiveName project.name + '.jar'
+	exclude '**/*.xml'
+	exclude '**/*.properties'
+	exclude '**/*.crt'
+	exclude '**/*.key'
 
     doLast {
 		copy {
@@ -65,13 +75,13 @@ jar {
 		}
 		copy {
 			from file('src/main/resources/applicationContext-sample.xml')
-      		from file('src/main/resources/log4j.properties')
+      from file('src/main/resources/log4j.properties')
+      from file('src/main/resources/privateKey.properties')
 			into 'dist/conf'
 		}
 		copy {
 			from file('tools/start.sh')
 			from file('tools/replace_solc_jar.sh')
-			from file('tools/get_account.sh')
 			into 'dist/'
 		}
 		copy {
