+++ src/main/java/console/common/Common.java
@@ -1,23 +1,18 @@
 package console.common;
 
 public class Common {
-
-    public static final String ContractLogFileName = "deploylog.txt";
-
-    // SystemConfig key
-    public static final String TxCountLimit = "tx_count_limit";
-    public static final String TxGasLimit = "tx_gas_limit";
-
-    public static final int InvalidReturnNumber = -100;
-    public static final int QueryLogCount = 20;
-    public static final int LogMaxCount = 10000;
-    public static final String PositiveIntegerRange = "from 1 to 2147483647";
-    public static final String NonNegativeIntegerRange = "from 0 to 2147483647";
-    public static final String DeployLogntegerRange = "from 1 to 100";
-    public static final String NodeIdLength = "128";
-    public static final String TxGasLimitRange = "from 100000 to 2147483647";
-    public static final int TxGasLimitMin = 10000;
-
-    public static int PermissionCode = 0;
-    public static final String PermissionDenied = "Permission denied.";
+	
+	// SystemConfig key
+	public static final String  TxCountLimit = "tx_count_limit";
+	public static final String TxGasLimit = "tx_gas_limit";
+	
+	public static final int InvalidReturnNumber = -100;
+	public static final int QueryLogCount = 20;
+	public static final int LogMaxCount = 10000;
+	public static final String PositiveIntegerRange = "from 1 to 2147483647";
+	public static final String NonNegativeIntegerRange = "from 0 to 2147483647";
+	public static final String DeployLogntegerRange = "from 1 to 100";
+	public static final String NodeIdLength = "128";
+	public static final String TxGasLimitRange = "from 100000 to 2147483647";
+	
 }
