+++ src/main/java/console/common/ConsoleVersion.java
@@ -2,9 +2,10 @@ package console.common;
 
 public class ConsoleVersion {
 
-    public static final String Version = "1.0.3";
+	public static final String Version = "1.0.2";
+	
+	public static void main(String[] args) {
+		System.out.println("console version: " + Version);
+	}
 
-    public static void main(String[] args) {
-        System.out.println("console version: " + Version);
-    }
 }
