--- src/main/java/console/common/EnumTransactionException.java
@@ -1,3 +1,0 @@
-package console.common;
-
-public enum EnumTransactionException {}
