+++ src/main/java/console/exception/ConsoleMessageException.java
@@ -4,21 +4,22 @@ import java.io.IOException;
 
 public class ConsoleMessageException extends IOException {
 
-    private static final long serialVersionUID = 1L;
+	private static final long serialVersionUID = 1L;
 
-    public ConsoleMessageException() {
-        super();
-    }
+	public ConsoleMessageException() {
+		super();
+	}
 
-    public ConsoleMessageException(String message, Throwable cause) {
-        super(message, cause);
-    }
+	public ConsoleMessageException(String message, Throwable cause) {
+		super(message, cause);
+	}
 
-    public ConsoleMessageException(String message) {
-        super(message);
-    }
+	public ConsoleMessageException(String message) {
+		super(message);
+	}
 
-    public ConsoleMessageException(Throwable cause) {
-        super(cause);
-    }
+	public ConsoleMessageException(Throwable cause) {
+		super(cause);
+	}
+	
 }
