+++ src/main/resources/applicationContext-sample.xml
@@ -33,7 +33,7 @@
 
 	<bean id="channelService" class="org.fisco.bcos.channel.client.Service" depends-on="groupChannelConnectionsConfig">
 		<property name="groupId" value="1" />
-		<property name="agencyName" value="fisco" />
+		<property name="orgID" value="fisco" />
 		<property name="allChannelConnections" ref="groupChannelConnectionsConfig"></property>
 	</bean>
