+++ src/main/resources/log4j.properties
@@ -1,49 +1,25 @@
 ### set log levels ###
-log4j.rootLogger = INFO, info, error
+#log4j.rootLogger = INFO , C , D , E
+log4j.rootLogger = DEBUG , D , E
 
-### output the TRACE level log information to the =./log/trace.log ###
-log4j.appender.trace = org.apache.log4j.DailyRollingFileAppender
-log4j.appender.trace.File = ./log/trace.log
-log4j.appender.trace.Append = true
-log4j.appender.trace.Threshold = TRACE
-log4j.appender.trace.filter.traceFilter = org.apache.log4j.varia.LevelRangeFilter
-log4j.appender.trace.filter.traceFilter.levelMin = TRACE
-log4j.appender.trace.filter.traceFilter.levelMax = TRACE
-log4j.appender.trace.layout = org.apache.log4j.PatternLayout
-log4j.appender.trace.layout.ConversionPattern = [%p] [%-d{yyyy-MM-dd HH:mm:ss}] %C{1}.%M(%L) | %m%n
+###output the log information to the console###
+log4j.appender.C = org.apache.log4j.ConsoleAppender
+log4j.appender.C.Target = System.out
+log4j.appender.C.layout = org.apache.log4j.PatternLayout
+log4j.appender.C.layout.ConversionPattern = [%p] [%-d{yyyy-MM-dd HH:mm:ss}] %C.%M(%L) | %m%n
 
-### output the DEBUG level log information to the =./log/debug.log ###
-log4j.appender.debug = org.apache.log4j.DailyRollingFileAppender
-log4j.appender.debug.File = ./log/debug.log
-log4j.appender.debug.Append = true
-log4j.appender.debug.Threshold = DEBUG
-log4j.appender.debug.filter.debugFilter = org.apache.log4j.varia.LevelRangeFilter
-log4j.appender.debug.filter.debugFilter.levelMin = DEBUG
-log4j.appender.debug.filter.debugFilter.levelMax = DEBUG
-log4j.appender.debug.layout = org.apache.log4j.PatternLayout
-log4j.appender.debug.layout.ConversionPattern = [%p] [%-d{yyyy-MM-dd HH:mm:ss}] %C{1}.%M(%L) | %m%n
+### output the INFO level log information to the =./log/demo_debug.log ###
+log4j.appender.D = org.apache.log4j.DailyRollingFileAppender
+log4j.appender.D.File = ./log/demo_debug.log
+log4j.appender.D.Append = true
+log4j.appender.D.Threshold = DEBUG 
+log4j.appender.D.layout = org.apache.log4j.PatternLayout
+log4j.appender.D.layout.ConversionPattern = [%p] [%-d{yyyy-MM-dd HH:mm:ss}] %C.%M(%L) | %m%n
 
-### output the INFO level log information to the =./log/info.log ###
-log4j.appender.info = org.apache.log4j.DailyRollingFileAppender
-log4j.appender.info.File = ./log/info.log
-log4j.appender.info.Append = true 
-log4j.appender.info.Threshold = INFO
-log4j.appender.info.filter.infoFilter = org.apache.log4j.varia.LevelRangeFilter
-log4j.appender.info.filter.infoFilter.levelMin = INFO
-log4j.appender.info.filter.infoFilter.levelMax = INFO
-log4j.appender.info.layout = org.apache.log4j.PatternLayout
-log4j.appender.info.layout.ConversionPattern = [%p] [%-d{yyyy-MM-dd HH:mm:ss}] %C{1}.%M(%L) | %m%n
-
-###  output the ERROR level log information to the=./log/error.log ###
-log4j.appender.error = org.apache.log4j.DailyRollingFileAppender
-log4j.appender.error.File = ./log/error.log 
-log4j.appender.error.Append = true
-log4j.appender.error.Threshold = ERROR
-log4j.appender.error.layout = org.apache.log4j.PatternLayout
-log4j.appender.error.layout.ConversionPattern = [%p] [%-d{yyyy-MM-dd HH:mm:ss}] %C{1}.%M(%L) | %m%n
-
-###output the log information to the console ###
-log4j.appender.stdout = org.apache.log4j.ConsoleAppender
-log4j.appender.stdout.Target = System.out
-log4j.appender.stdout.layout = org.apache.log4j.PatternLayout
-log4j.appender.stdout.layout.ConversionPattern = [%p] [%-d{yyyy-MM-dd HH:mm:ss}] %C{1}.%M(%L) | %m%n
+###  output the  ERROR level log information to the=./log/demo_error.log ###
+log4j.appender.E = org.apache.log4j.DailyRollingFileAppender
+log4j.appender.E.File = ./log/demo_error.log 
+log4j.appender.E.Append = true
+log4j.appender.E.Threshold = ERROR 
+log4j.appender.E.layout = org.apache.log4j.PatternLayout
+log4j.appender.E.layout.ConversionPattern = [%p] [%-d{yyyy-MM-dd HH:mm:ss}] %C.%M(%L) | %m%n
