+++ tools/download_console.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
-version=$(curl -s https://api.github.com/repos/FISCO-BCOS/console/releases | grep "tag_name" | sort -u | tail -n 1 | cut -d \" -f 4 | sed "s/^[vV]//")
-echo "Downloading console version" ${version}
+
+version=$(curl -s https://raw.githubusercontent.com/FISCO-BCOS/console/master/release_note.txt | sed "s/^[vV]//")
 package_name="console.tar.gz"
 download_link=https://github.com/FISCO-BCOS/console/releases/download/v${version}/${package_name}
 curl -LO ${download_link}
